third_party/blink/web_tests/external/wpt/html/browsers/sandboxing/sandbox-inherited-from-required-csp-expected.txt

This is a testharness.js-based test.
[FAIL] initial empty document
  assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
[PASS] about:blank
[PASS] data-url
[PASS] srcdoc
[PASS] blob URL
[PASS] same-origin
[PASS] cross-origin
Harness: the test ran to completion.
